+++ client/src/components/InfoBar.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import onlineIcon from '../icons/onlineIcon.png';
 import closeIcon from '../icons/closeIcon.png';
 
-import './InfoBar.scss';
+import '../styles/InfoBar.scss';
 
 const InfoBar = ({room}) => (
   <div className="infoBar">
