+++ client/src/styles/InfoBar.scss
