+++ client/src/styles/Join.scss
