+++ client/src/styles/chatButton.scss
